fix(wgpu): clamp max_n_sub in u64 space + correct scores_buf header comment
Address PR #270 review:
- max_n_sub cast to u32 happened before .min(n); a binding limit large
  enough that the division exceeds u32::MAX could truncate mod 2^32 to a
  bogus (possibly 0) count, giving n_sub = 0 and an infinite tiling loop.
  Clamp in u64 space via .min(u64::from(n)) before the cast, which is then
  exact since the value is <= n.
- attention_prefill.wgsl bind-group header still described scores_buf as
  n_queries x n_heads x max_seq; with query tiling it is the dispatch-local
  n_sub_queries x n_heads x max_seq, addressed by q_idx. Comment corrected.

+++ cera/src/model/gpu_lfm2.rs
@@ -3763,8 +3763,11 @@ impl GpuLfm2Model {
         );
         // The assert above gives `per_query_scores ∈ (0, max_binding_floats]`, so
         // the division is ≥ 1; `.min(n)` then bounds it to the batch (and yields 0
-        // only when n == 0, where the loop below simply never runs).
-        let max_n_sub = ((max_binding_floats / per_query_scores) as u32).min(n);
+        // only when n == 0, where the loop below simply never runs). The `.min` is
+        // done in `u64` space *before* the `as u32` cast so a large binding limit
+        // (division result > u32::MAX) can't truncate to a bogus — possibly 0 —
+        // count; after `.min(u64::from(n))` the value is ≤ n and the cast is exact.
+        let max_n_sub = (max_binding_floats / per_query_scores).min(u64::from(n)) as u32;
 
         let mut q_base = 0u32;
         while q_base < n {
